Working with dict comprehensions for unpacking. #PyDev-573: Code Completion of compound types in PyDev 4.0

+++ plugins/org.python.pydev/src_completions/org/python/pydev/editor/codecompletion/revisited/CompletionState.java
@@ -66,7 +66,7 @@ public final class CompletionState implements ICompletionState {
     private ICompletionCache completionCache;
     private String fullActivationToken;
     private long initialMillis = 0;
-    private int maxMillisToComplete;
+    private long maxMillisToComplete;
 
     public ICompletionState getCopy() {
         return new CompletionStateWrapper(this);
